#2706 moved authentication logic inside synctemphomefiles method

+++ application/controllers/webservice/newhomewebservice.php
@@ -59,10 +59,6 @@ class NewHomeWebService extends MY_Controller
                                                                               $this->input->get('hash'),
                                                                               true);
         }
-        
-        if(!$this->isAuthenticated) {
-                redirect("/","refresh");
-        }
     }
 
     /**
@@ -1265,19 +1261,23 @@ class NewHomeWebService extends MY_Controller
      */
     public function syncTempHomeFiles()
     {
-        $map = simplexml_load_file($this->file);
+        if($this->isAuthenticated) {
+            $map = simplexml_load_file($this->file);
 
-        foreach ($map->sliderSection->slide as $slider) {
-            $sliders[] = $slider;
+            foreach ($map->sliderSection->slide as $slider) {
+                $sliders[] = $slider;
+            }
+
+            $this->xmlCmsService->removeXmlNode($this->tempHomefile, "tempHomeSlider");
+            $this->xmlCmsService->syncTempSliderValues($this->tempHomefile, $this->file, $sliders);  
+             
+            return $this->output
+                        ->set_content_type('application/json')
+                        ->set_output($this->json);                        
         }
-
-        $this->xmlCmsService->removeXmlNode($this->tempHomefile, "tempHomeSlider");
-        $this->xmlCmsService->syncTempSliderValues($this->tempHomefile, $this->file, $sliders);  
-         
-        return $this->output
-                    ->set_content_type('application/json')
-                    ->set_output($this->json);                        
-
+        else {
+            return json_encode("error");
+        }        
           
     }
